refactor(workflow): rename steps/script.rb to steps/ruby_script.rb
The file defines the RubyScript class. Name the file after the class
so future autoloaders do not choke and the path matches the 1:1
file/class convention the rest of lib/ follows.
